Add an error handler to the HEAD request
If you have an asset bundle already in your cache and on your next
request anything goes wrong with the head request, the failed request
does not get back to unity. In our application it meant that our requests
would not be retried and would not present error information to the
user.

This simply forwards the error back to the original request made from
within unity.